Middleware is always giving a string to the Shield::verify method
The Shield::verify method expects two strings, but $request->getUser and $request->getPassword are returning null if there is no user in the HTTP header of the request. So the verify method fails before it can throw the UnauthorizedHttpException('Basic'). But this is needed to get the login screen of the browser and send the appropriate HTTP Header with username and password.